// shelfy-reader-rn/src/types/reader.ts
export interface Chapter {
  id: string;
  label: string;
  href: string;
  subitems?: Chapter[];
}